v2.1.0 #25
NOTE TO DEVS:
1) wdl.patch has been updated for OSX-only swell-ini.cpp.  Patch if you build OSX, please.
2) reaper_plugin_functions.h no longer needs to be modified after it's generated in Reaper.

Also included are some x64 compiler warning fixes.

+++ Misc/Analysis.cpp
@@ -86,7 +86,7 @@ void AnalyzePCMSource(ANALYZE_PCM* a)
 	a->dProgress = 0.0;
 	a->sampleCount = 0;
 
-	INT64 totalSamples = a->pcm->GetLength() * t.samplerate;
+	INT64 totalSamples = (INT64)(a->pcm->GetLength() * t.samplerate);
 	int iFrame = 0;
 
 	a->pcm->GetSamples(&t);
